# snake-game-perro/openapi.yaml
openapi: 3.0.0
info:
  title: Snake Arena API
  version: 1.0.0
  description: API for the Snake Arena game
servers:
  - url: http://localhost:8000
    description: Local development server

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
      required:
        - id
        - username
        - email

    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
        error:
          type: string
      required:
        - success

    LeaderboardEntry:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        score:
          type: integer
        mode:
          type: string
          enum: [walls, pass-through]
        date:
          type: string
          format: date
      required:
        - id
        - username
        - score
        - mode
        - date

    GameSession:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        score:
          type: integer
        mode:
          type: string
          enum: [walls, pass-through]
        isLive:
          type: boolean
      required:
        - id
        - username
        - score
        - mode
        - isLive

paths:
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /auth/signup:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                username:
                  type: string
                password:
                  type: string
              required:
                - email
                - username
                - password
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /auth/logout:
    post:
      summary: Logout user
      responses:
        '200':
          description: Logout successful

  /auth/me:
    get:
      summary: Get current authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated

  /leaderboard:
    get:
      summary: Get leaderboard entries
      parameters:
        - in: query
          name: mode
          schema:
            type: string
            enum: [walls, pass-through]
      responses:
        '200':
          description: List of leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
    post:
      summary: Submit a score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                score:
                  type: integer
                mode:
                  type: string
                  enum: [walls, pass-through]
              required:
                - score
                - mode
      responses:
        '200':
          description: Score submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  rank:
                    type: integer
                required:
                  - success

  /sessions:
    get:
      summary: Get active game sessions
      responses:
        '200':
          description: List of active sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameSession'

  /sessions/{id}:
    get:
      summary: Get a specific session
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSession'
        '404':
          description: Session not found